pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "context-helpers"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = []

[project.optional-dependencies]
# Core server (always required)
server = ["fastapi", "uvicorn[standard]", "pydantic-settings>=2", "PyYAML", "click", "jinja2"]

# Per-collector extras (only install what you need)
reminders = []          # stdlib only: subprocess + osascript
imessage  = []          # stdlib only: sqlite3
music     = []          # stdlib only: plistlib + xml.etree
health    = ["healthkit-to-sqlite"]
notes     = ["apple-notes-to-sqlite"]

# Convenience bundle
all = ["healthkit-to-sqlite", "apple-notes-to-sqlite"]

dev = [
    "pytest",
    "httpx",
    "ruff",
    "mypy",
    # needed to import server/auth/cli without installing extras
    "fastapi",
    "uvicorn[standard]",
    "pydantic-settings>=2",
    "PyYAML",
    "click",
    "jinja2",
]

[project.scripts]
context-helpers = "context_helpers.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/context_helpers"]

[tool.pytest.ini_options]
pythonpath = ["src", "."]   # "." makes `tests/` importable as a package (for conftest imports)
testpaths = ["tests"]
